improve the menu
