+++ Cinema Reservation/main/view_controller_manager.py
@@ -2,6 +2,7 @@ from .controllers import UserController, TextController, TutorialController, Spe
 from .models import User
 from .views import (UserViews, MovieViews, ProjectionViews,
                    ReservationViews, system_input)
+from .urwid_views import Menu
 from views_constants import PROJECTION_SEATS
 from sqlalchemy.exc import IntegrityError
 import re
@@ -23,22 +24,17 @@ class ViewControllerManager:
         # self.reservation_controllers = ReservationController()
 
     def manage_entering_system_views_and_controllers(self):
-        is_system_entered = False
         user_entered_system = None
-        while not is_system_entered:
-            command = self.user_views.console_read_command_view()
-            if command == 'help':
-                self.user_views.guest_user_help_view()
-            elif command == 'login':
-                user_entered_system = self.manage_login_view_and_controller()
-            elif command == 'signup':
-                user_entered_system = self.manage_signup_view_and_controller()
-            elif command == 'exit':
-                raise SystemExit
-            else:
-                print(f'\'{command}\' command not found')
-            if user_entered_system is not None:
-                is_system_entered = True
+        menu = Menu('welcome', ['login', 'signup', 'exit'])
+        command = menu.command
+        if command == 'login':
+            user_entered_system = self.manage_login_view_and_controller()
+        elif command == 'signup':
+            user_entered_system = self.manage_signup_view_and_controller()
+        elif command == 'exit':
+            raise SystemExit
+        else:
+            raise ValueError
 
         return user_entered_system
 
@@ -181,6 +177,6 @@ class ViewControllerManager:
 
     def release_resources(self):
         self.user_controllers.gateway.db.close()
-        self.movie_controllers.gateway.db.close()
-        self.projection_controllers.gateway.db.close()
-        self.reservation_controllers.gateway.db.close()
+        self.text_controllers.gateway.db.close()
+        self.tutorial_controllers.gateway.db.close()
+        self.speedtest_controllers.gateway.db.close()
