+++ Cinema Reservation/main2.py
@@ -1,16 +1,10 @@
-import os
 from setup_database import setup_database
 from main.views import clear_screen
 from main.view_controller_manager import ViewControllerManager
 
 
-def welcome():
-    print('----- WELCOME to our cinema reservation system! -----'.center(os.get_terminal_size().columns))
-
-
 def start():
     clear_screen()
-    welcome()
     manager = ViewControllerManager()
     try:
         user = manager.manage_entering_system_views_and_controllers()
