--- Touch-typing-tutorial/Cinema Reservation/main/__init__.py
@@ -1,1 +1,0 @@
-from main.utils import validate_email, validate_password
