--- Touch-typing-tutorial/Cinema Reservation/rebuild.sh
@@ -1,3 +1,0 @@
-rm cinema.db
-touch cinema.db
-python main.py
