--- Touch-typing-tutorial/Cinema Reservation/requirements.txt
@@ -1,4 +1,0 @@
-termcolor==1.1.0
-tabulate==0.8.7
-SQLAlchemy==1.3.16
-
