--- Touch-typing-tutorial/Cinema Reservation/settings.py
@@ -1,4 +1,0 @@
-DB_NAME = 'cinema.db'
-SU_NAME = 'admin'
-SU_PASS = 'admin'
-HASHING_TIMES = 100
