--- Touch-typing-tutorial/Cinema Reservation/views_constants.py
@@ -1,12 +1,0 @@
-INPUT_COLOR = 'blue'
-COLOR_IN_EXIT = 'red'
-ID_COLOR = 'green'
-COMMAND_COLOR = 'cyan'
-
-INVALID_OPTION_TEXT_COLOR = 'red'
-VALID_LOGIN_OR_SIGNUP_RETURNED_VALUE = 0
-INVALID_LOGIN_RETURNED_VALUE = 1
-INVALID_SIGHNUP_RETURNED_VALUE = 2
-INVALID_CHOICE_RETURNED_VALUE = 3
-
-PROJECTION_SEATS = 25
