--- Touch-typing-tutorial/README.md
@@ -1,2 +1,0 @@
-# Touch-typing-tutorial
-a tutorial of typing with ten fingers
